+++ setup.py
@@ -22,7 +22,7 @@ from setuptools import setup, find_packages
 
 # Parse version out of descarteslabs/__init__.py
 _version_re = re.compile(r'__version__\s+=\s+(.*)')
-with open('descarteslabs/__init__.py', 'rb') as f:
+with open('descarteslabs/version.py', 'rb') as f:
     version = str(ast.literal_eval(_version_re.search(
         f.read().decode('utf-8')).group(1))
     )
